Code cleanup and organization
